Split tests and convert all services to use test data
